Added explosion animation

+++ AnimationData.cpp
@@ -2,6 +2,6 @@
 
 ani::AnimationData::AnimationData() {}
 
-ani::AnimationData::AnimationData(int i, int f, int s) 
+ani::AnimationData::AnimationData(int i, int f, float s) 
 
     :index(i), frames(f), speed(s) {}
